Struct epic::tui::table::TableView

pub struct TableView<T: TableViewItem<H>, H: Eq + Hash + Copy + Clone + 'static> {
    enabled: bool,
    scrollbase: ScrollBase,
    last_size: Vec2,
    column_select: bool,
    columns: Vec<TableColumn<H>>,
    column_indices: HashMap<H, usize>,
    focus: usize,
    items: Vec<T>,
    rows_to_items: Vec<usize>,
    on_sort: Option<Rc<dyn Fn(&mut Cursive, H, Ordering)>>,
    on_submit: Option<Rc<dyn Fn(&mut Cursive, usize, usize)>>,
    on_select: Option<Rc<dyn Fn(&mut Cursive, usize, usize)>>,
}

View to select an item among a list, supporting multiple columns for sorting.

Examples

// Provide a type for the table's columns
#[derive(Copy, Clone, PartialEq, Eq, Hash)]
enum BasicColumn {
    Name,
    Count,
    Rate
}

// Define the item type
#[derive(Clone, Debug)]
struct Foo {
    name: String,
    count: usize,
    rate: usize
}

impl TableViewItem<BasicColumn> for Foo {

    fn to_column(&self, column: BasicColumn) -> String {
        match column {
            BasicColumn::Name => self.name.to_string(),
            BasicColumn::Count => format!("{}", self.count),
            BasicColumn::Rate => format!("{}", self.rate)
        }
    }

    fn cmp(&self, other: &Self, column: BasicColumn) -> Ordering where Self: Sized {
        match column {
            BasicColumn::Name => self.name.cmp(&other.name),
            BasicColumn::Count => self.count.cmp(&other.count),
            BasicColumn::Rate => self.rate.cmp(&other.rate)
        }
    }

}

// Configure the actual table
let table = TableView::<Foo, BasicColumn>::new()
                     .column(BasicColumn::Name, "Name", |c| c.width(20))
                     .column(BasicColumn::Count, "Count", |c| c.align(HAlign::Center))
                     .column(BasicColumn::Rate, "Rate", |c| {
                         c.ordering(Ordering::Greater).align(HAlign::Right).width(20)
                     })
                     .default_column(BasicColumn::Name);

Fields
enabled: boolscrollbase: ScrollBaselast_size: Vec2column_select: boolcolumns: Vec<TableColumn<H>>column_indices: HashMap<H, usize>focus: usizeitems: Vec<T>rows_to_items: Vec<usize>on_sort: Option<Rc<dyn Fn(&mut Cursive, H, Ordering)>>on_submit: Option<Rc<dyn Fn(&mut Cursive, usize, usize)>>on_select: Option<Rc<dyn Fn(&mut Cursive, usize, usize)>>
Implementations

impl<T: TableViewItem<H>, H: Eq + Hash + Copy + Clone + 'static> TableView<T, H>

pub fn new() -> Self

Creates a new empty TableView without any columns.

A TableView should be accompanied by a enum of type H representing the table columns.

pub fn column<S: Into<String>, C: FnOnce(TableColumn<H>) -> TableColumn<H>>(
self,
column: H,
title: S,
callback: C
) -> Self

Adds a column for the specified table column from type H along with a title for its visual display.

The provided callback can be used to further configure the created TableColumn.

pub fn default_column(self, column: H) -> Self

Sets the initially active column of the table.

pub fn sort_by(&mut self, column: H, order: Ordering)

Sorts the table using the specified table column and the passed order.

pub fn sort(&mut self)

Sorts the table using the currently active column and its ordering.

pub fn order(&self) -> Option<(H, Ordering)>

Returns the currently active column that is used for sorting along with its ordering.

Might return None if there are currently no items in the table and it has not been sorted yet.

pub fn disable(&mut self)

Disables this view.

A disabled view cannot be selected.

pub fn enable(&mut self)

Re-enables this view.

pub fn set_enabled(&mut self, enabled: bool)

Enable or disable this view.

pub fn is_enabled(&self) -> bool

Returns true if this view is enabled.

pub fn set_on_sort<F>(&mut self, cb: F) where
F: Fn(&mut Cursive, H, Ordering) + 'static,

Sets a callback to be used when a selected column is sorted by pressing <Enter>.

Example

rt(|siv: &mut Cursive, column: BasicColumn, order: Ordering| {});

pub fn on_sort<F>(self, cb: F) -> Self where
F: Fn(&mut Cursive, H, Ordering) + 'static,

Sets a callback to be used when a selected column is sorted by pressing <Enter>.

Chainable variant.

Example

siv: &mut Cursive, column: BasicColumn, order: Ordering| {});

pub fn set_on_submit<F>(&mut self, cb: F) where
F: Fn(&mut Cursive, usize, usize) + 'static,

Sets a callback to be used when <Enter> is pressed while an item is selected.

Both the currently selected row and the index of the corresponding item within the underlying storage vector will be given to the callback.

Example

bmit(|siv: &mut Cursive, row: usize, index: usize| {});

pub fn on_submit<F>(self, cb: F) -> Self where
F: Fn(&mut Cursive, usize, usize) + 'static,

Sets a callback to be used when <Enter> is pressed while an item is selected.

Both the currently selected row and the index of the corresponding item within the underlying storage vector will be given to the callback.

Chainable variant.

Example

(|siv: &mut Cursive, row: usize, index: usize| {});

pub fn set_on_select<F>(&mut self, cb: F) where
F: Fn(&mut Cursive, usize, usize) + 'static,

Sets a callback to be used when an item is selected.

Both the currently selected row and the index of the corresponding item within the underlying storage vector will be given to the callback.

Example

lect(|siv: &mut Cursive, row: usize, index: usize| {});

pub fn on_select<F>(self, cb: F) -> Self where
F: Fn(&mut Cursive, usize, usize) + 'static,

Sets a callback to be used when an item is selected.

Both the currently selected row and the index of the corresponding item within the underlying storage vector will be given to the callback.

Chainable variant.

Example

(|siv: &mut Cursive, row: usize, index: usize| {});

pub fn clear(&mut self)

Removes all items from this view.

pub fn len(&self) -> usize

Returns the number of items in this table.

pub fn is_empty(&self) -> bool

Returns true if this table has no items.

pub fn row(&self) -> Option<usize>

Returns the index of the currently selected table row.

pub fn set_selected_row(&mut self, row_index: usize)

Selects the row at the specified index.

pub fn selected_row(self, row_index: usize) -> Self

Selects the row at the specified index.

Chainable variant.

pub fn set_items(&mut self, items: Vec<T>)

Sets the contained items of the table.

The currently active sort order is preserved and will be applied to all items.

pub fn items(self, items: Vec<T>) -> Self

Sets the contained items of the table.

The order of the items will be preserved even when the table is sorted.

Chainable variant.

pub fn borrow_item(&mut self, index: usize) -> Option<&T>

Returns a immutable reference to the item at the specified index within the underlying storage vector.

pub fn borrow_item_mut(&mut self, index: usize) -> Option<&mut T>

Returns a mutable reference to the item at the specified index within the underlying storage vector.

pub fn borrow_items(&mut self) -> &Vec<T>

Returns a immutable reference to the items contained within the table.

pub fn borrow_items_mut(&mut self) -> &mut Vec<T>

Returns a mutable reference to the items contained within the table.

Can be used to modify the items in place.

pub fn item(&self) -> Option<usize>

Returns the index of the currently selected item within the underlying storage vector.

pub fn set_selected_item(&mut self, item_index: usize)

Selects the item at the specified index within the underlying storage vector.

pub fn selected_item(self, item_index: usize) -> Self

Selects the item at the specified index within the underlying storage vector.

Chainable variant.

pub fn insert_item(&mut self, item: T)

Inserts a new item into the table.

The currently active sort order is preserved and will be applied to the newly inserted item.

pub fn remove_item(&mut self, item_index: usize) -> Option<T>

Removes the item at the specified index within the underlying storage vector and returns it.

pub fn take_items(&mut self) -> Vec<T>

Removes all items from the underlying storage and returns them.

impl<T: TableViewItem<H>, H: Eq + Hash + Copy + Clone + 'static> TableView<T, H>

fn draw_columns<C: Fn(&Printer<'_, '_>, &TableColumn<H>)>(
&self,
printer: &Printer<'_, '_>,
sep: &str,
callback: C
)

fn sort_items(&mut self, column: H, order: Ordering)

fn draw_item(&self, printer: &Printer<'_, '_>, i: usize)

fn focus_up(&mut self, n: usize)

fn focus_down(&mut self, n: usize)

fn active_column(&self) -> usize

fn column_cancel(&mut self)

fn column_next(&mut self) -> bool

fn column_prev(&mut self) -> bool

fn column_select(&mut self)