Trait epic::tui::table::TableViewItem

pub trait TableViewItem<H>: Clone + Sized where
    H: Eq + Hash + Copy + Clone + 'static, {
    fn to_column(&self, column: H) -> String;

    fn cmp(&self, other: &Self, column: H) -> Ordering
    where
        Self: Sized;
}

A trait for displaying and sorting items inside a TableView.

Required methods

fn to_column(&self, column: H) -> String

Method returning a string representation of the item for the specified column from type H.

fn cmp(&self, other: &Self, column: H) -> Ordering where
Self: Sized,

Method comparing two items via their specified column from type H.